feat: created login/logout button and imported to header
