+++ src/components/Header/Header.component.jsx
@@ -1,14 +1,13 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { Link } from 'react-router-dom';
-import { Col, Header, LogIcon, LoginContainer, MenuIcon, Row } from './Header.styles';
-import { ToggleSwitch } from '../atoms';
+import { Col, Header, LogInIcon, LogOutIcon, LoginContainer, Row } from './Header.styles';
+import { ToggleSwitch, LogButton } from '../atoms';
 import { SearchBar, SideBar } from '../molecules';
 
 const MainHeader = () => {
-  const [isToggled, setIsToggled] = useState(false);
 
   return (
-    <Header darkMode={isToggled}>
+    <Header>
       <Row>
         <Col size={1}>
           <SideBar />
@@ -18,17 +17,10 @@ const MainHeader = () => {
         </Col>
         <Col size={1}>
           <ToggleSwitch
-            id="test-switch"
-            toggled={isToggled}
-            onChange={(e) => setIsToggled(e.target.checked)}
           />
         </Col>
         <Col size={1}>
-          <LoginContainer>
-            <Link to="/login">
-              Login <LogIcon />
-            </Link>
-          </LoginContainer>
+          <LogButton />
         </Col>
       </Row>
     </Header>
