+++ src/components/Header/Header.styles.js
@@ -1,5 +1,4 @@
 import styled from 'styled-components';
-import { VscSymbolBoolean } from 'react-icons/vsc';
 
 export const Header = styled.header`
   display: flex;
@@ -20,23 +19,4 @@ export const Row = styled.div`
 
 export const Col = styled.div`
   flex: ${(props) => props.size};
-`;
-
-export const LoginContainer = styled.div`
-  display: block;
-  flex-direction: row;
-  align-text: center;
-`;
-// color: #F894F1; pink
-
-export const LogIcon = styled(VscSymbolBoolean)`
-  font-size: 27px;
-  margin-right: 20px;
-  color: #6c56a3;
-  border: none;
-  transform: scaleX(-1);
-  -moz-transform: scaleX(-1);
-  -webkit-transform: scaleX(-1);
-  -ms-transform: scaleX(-1);
-  margin-left: 0.2em;
 `;
