+++ src/components/atoms/index.js
@@ -2,5 +2,6 @@ import ToggleSwitch from './ToggleSwitch/ToggleSwitch.atoms';
 import VideoInformation from './VideoInfo/VideoInfo.atom';
 import Avatar from './Avatar/Avatar.atom';
 import Loader from './Loader/Loader.atom';
+import LogButton from './LogButton/LogButton.atom';
 
-export { ToggleSwitch, VideoInformation, Avatar, Loader };
+export { ToggleSwitch, VideoInformation, Avatar, Loader, LogButton };
