+++ src/pages/Home/Home.page.jsx
@@ -1,19 +1,8 @@
-import React, { useRef } from 'react';
-import { Link, useHistory } from 'react-router-dom';
-import { useAuth } from '../../providers/Auth';
+import React from 'react';
 import YoutubeVideos from '../../components/YoutubeVideos/YoutubeVideos.component';
 import { Home, HomeTitle, VideoListSection } from './Home.styles.js';
 
 const HomePage = () => {
-  const history = useHistory();
-  const sectionRef = useRef(null);
-  const { authenticated, logout } = useAuth();
-
-  function deAuthenticate(event) {
-    event.preventDefault();
-    logout();
-    history.push('/');
-  }
 
   return (
     <>
@@ -28,21 +17,3 @@ const HomePage = () => {
 
 export default HomePage;
 
-
-    // <section className="homepage" ref={sectionRef}>
-    //   <h1>Hello stranger!</h1>
-    //   {authenticated ? (
-    //     <>
-    //       <h2>Good to have you back</h2>
-    //       <span>
-    //         <Link to="/" onClick={deAuthenticate}>
-    //           ← logout
-    //         </Link>
-    //         <span className="separator" />
-    //         <Link to="/secret">show me something cool →</Link>
-    //       </span>
-    //     </>
-    //   ) : (
-    //     <Link to="/login">let me in →</Link>
-    //   )}
-    // </section>
