+++ src/pages/Login/Login.page.jsx
@@ -16,8 +16,8 @@ function LoginPage() {
 
   return (
     <section className="login">
-      <h1>HALLELLO!</h1>
-      <h5>You can log in now, you know why?</h5>
+      <h1>HaLLeLoO!</h1>
+      <h5>You can login, you know why?</h5>
       <h3>Because you are Sickening!</h3>
       <form onSubmit={authenticate} className="login-form">
         <div className="form-group">
