+++ src/pages/Login/Login.styles.css
@@ -11,8 +11,7 @@
 .login h3, h5 {
   text-align: center;
   letter-spacing: -1px;
-  
-  color: #D4AF37
+  color: #D4AF37;
 }
 
 .login-form {
@@ -33,6 +32,7 @@
   font-weight: 700;
   text-transform: capitalize;
   margin-bottom: 0.4rem;
+  color: #432272;
 }
 
 .form-group input {
